+++ src/main/kotlin/com/github/aivancioglo/resttest/http/Model.kt
@@ -4,6 +4,10 @@ import com.github.aivancioglo.resttest.logger.Logger
 import io.restassured.internal.RequestSpecificationImpl
 
 abstract class Model : Response() {
+
+    /**
+     * This method can be used for setting of all required variables of default response.
+     */
     fun set(request: RequestSpecificationImpl, response: io.restassured.response.Response) {
         super.request = request
         super.response = response