+++ changeLog.md
@@ -1,3 +1,14 @@
+## Next planned version
+##### Features:
+* New properties: `protocol`, `host`, `decoder_charset` and `encoder_charset`.
+* New content type: `MULTIPART`. 
+
+## V. 1.0.3
+##### Fixes:
+* Fixed issue when `log_if_failed` is `true` and `soft_assertions` is `false` but passive log is not working.
+* Fixed issue when `soft_assertions` is `false` but all asserts are performing.
+* Fix issue when all user properties are setting after request. 
+
 ## V. 1.0.2
 ##### Fixes:
 * Fixed null pointer exception when `Content-Lenght` header is `null`.